requirements.txt
-f /cvmfs/soft.computecanada.ca/custom/python/wheelhouse/avx2
-f /cvmfs/soft.computecanada.ca/custom/python/wheelhouse/generic
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.16.2
Pillow==5.3.0
pyparsing==2.3.1
python-dateutil==2.8.0
scipy==1.2.1
six==1.12.0
torch-gpu==1.0.0
torchvision==0.2.1
